+++ package.json
@@ -13,6 +13,10 @@
     "deploy-storybook": "cd storybook && npm run deploy-storybook"
   },
   "devDependencies": {
+    "@babel/core": "^7.11.6",
+    "@babel/preset-env": "^7.11.5",
+    "@babel/preset-react": "^7.10.4",
+    "@babel/preset-typescript": "^7.10.4",
     "@types/jest": "^26.0.10",
     "@zouloux/files": "^0.6.0",
     "app-root-path": "^3.0.0",
