+++ packages/react-hooks/use-async-effect/src/index.ts
@@ -1,0 +1,2 @@
+export { useAsyncEffect } from "./useAsyncEffect";
+export { useAsyncLayoutEffect } from "./useAsyncLayoutEffect";
