+++ packages/react-hooks/use-async-effect/src/useAsyncEffect.tsx
@@ -1,0 +1,15 @@
+import { useEffect } from "react";
+
+const componentName: string = "useAsyncEffect";
+const debug = require("debug")(`lib:${componentName}`);
+
+/**
+ * useAsyncEffect
+ */
+function useAsyncEffect(pEffect: () => void, pInputs?: any[]): void {
+  useEffect(() => {
+    pEffect();
+  }, pInputs);
+}
+
+export { useAsyncEffect };
