+++ packages/react-hooks/use-async-effect/src/useAsyncLayoutEffect.tsx
@@ -1,0 +1,15 @@
+import { useLayoutEffect } from "react";
+
+const componentName: string = "useAsyncLayoutEffect";
+const debug = require("debug")(`lib:${componentName}`);
+
+/**
+ * useAsyncLayoutEffect
+ */
+function useAsyncLayoutEffect(pEffect: () => void, pInputs?: any[]): void {
+  useLayoutEffect(() => {
+    pEffect();
+  }, pInputs);
+}
+
+export { useAsyncLayoutEffect };
