+++ packages/react-hooks/use-async-effect/test/useAsyncEffect.test.ts
@@ -1,0 +1,8 @@
+// @ts-ignore
+import { useAsyncEffect } from "../src";
+
+describe("useAsyncEffect", () => {
+  it("useAsyncEffect should be defined", () => {
+    expect(useAsyncEffect).toBeDefined();
+  });
+});
