+++ packages/react-hooks/use-async-effect/test/useAsyncLayoutEffect.test.ts
@@ -1,0 +1,8 @@
+// @ts-ignore
+import { useAsyncLayoutEffect } from "../src";
+
+describe("useAsyncLayoutEffect", () => {
+  it("useAsyncLayoutEffect should be defined", () => {
+    expect(useAsyncLayoutEffect).toBeDefined();
+  });
+});
