+++ packages/react-hooks/use-bounding-client-rect/test/useBoundingClientRect.test.ts
@@ -1,0 +1,8 @@
+// @ts-ignore
+import useBoundingClientRect from "../src";
+
+describe("useBoundingClientRect", () => {
+  it("should be defined", () => {
+    expect(useBoundingClientRect).toBeDefined();
+  });
+});
