+++ packages/react-hooks/use-bounding-client-rect/tsconfig.json
@@ -4,5 +4,5 @@
     "outDir": "./dist",
     "rootDir": "./src"
   },
-  "exclude": ["node_modules", "dist", "example"]
+  "exclude": ["node_modules", "dist", "test"]
 }
