+++ packages/react-hooks/use-document-scroll-top/test/useDocumentScrollTop.test.ts
@@ -1,0 +1,8 @@
+// @ts-ignore
+import useDocumentScrollTop from "../src";
+
+describe("useDocumentScrollTop", () => {
+  it("should be defined", () => {
+    expect(useDocumentScrollTop).toBeDefined();
+  });
+});
