+++ packages/react-hooks/use-document-scroll-top/tsconfig.json
@@ -4,5 +4,6 @@
     "outDir": "./dist",
     "rootDir": "./src"
   },
-  "exclude": ["node_modules", "dist"]
+  "exclude": ["node_modules", "dist", "test"],
+  "include": ["src/*"]
 }
