+++ packages/react-hooks/use-is-in-viewport/.babelrc
@@ -1,0 +1,11 @@
+{
+  "presets": [
+    ["@babel/preset-env", {
+      "targets": {
+        "node": "current"
+      }
+    }],
+    "@babel/preset-typescript",
+    "@babel/preset-react"
+  ]
+}
