+++ packages/react-hooks/use-is-in-viewport/package.json
@@ -14,7 +14,8 @@
   "scripts": {
     "build": "npm run clean && npm run compile",
     "clean": "rm -rf ./dist && rm -rf tsconfig.tsbuildinfo",
-    "compile": "tsc -b tsconfig.json"
+    "compile": "tsc -b tsconfig.json",
+    "test": "jest"
   },
   "dependencies": {
     "@wbe/use-bounding-client-rect": "^1.3.0",
