+++ packages/react-hooks/use-is-in-viewport/test/useIsInViewport.test.ts
@@ -1,0 +1,8 @@
+// @ts-ignore
+import useIsInViewport from "../src";
+
+describe("useIsInViewport", () => {
+  it("should be defined", () => {
+    expect(useIsInViewport).toBeDefined();
+  });
+});
