+++ packages/react-hooks/use-responsive-image-data/package.json
@@ -14,7 +14,8 @@
   "scripts": {
     "build": "npm run clean && npm run compile",
     "clean": "rm -rf ./dist && rm -rf tsconfig.tsbuildinfo",
-    "compile": "tsc -b tsconfig.json"
+    "compile": "tsc -b tsconfig.json",
+    "test": "jest"
   },
   "dependencies": {
     "@wbe/use-window-size": "^1.3.1-alpha.1",
