+++ packages/react-hooks/use-responsive-image-data/test/useResponsiveImageData.test.ts
@@ -1,0 +1,8 @@
+// @ts-ignore
+import useResponsiveImageData from "../src";
+
+describe("useResponsiveImageData", () => {
+  it("should be defined", () => {
+    expect(useResponsiveImageData).toBeDefined();
+  });
+});
