+++ packages/react-hooks/use-window-size/package.json
@@ -19,10 +19,6 @@
     "react": "^16.13.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.4",
-    "@babel/preset-env": "^7.11.0",
-    "@babel/preset-react": "^7.10.4",
-    "@babel/preset-typescript": "^7.10.4",
     "@types/react": "^16.9.48"
   },
   "keywords": [
